insert type casting to quiet msvc compiler warnings. insert header files for Win32 builds for similar reasons.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V5-1-patches@9439 06827809-a52a-0410-b366-d66718629ded

+++ agent/helpers/table_iterator.c
@@ -479,7 +479,7 @@ netsnmp_table_iterator_helper_handler(netsnmp_mib_handler *handler,
 
                         table_info->indexes = index_search;
                         for(i = table_reg_info->min_column;
-                            i <= table_reg_info->max_column; i++) {
+                            i <= (int)table_reg_info->max_column; i++) {
                             myname[reginfo->rootoid_len + 1] = i;
                             table_info->colnum = i;
                             vb = reqtmp->requestvb =
